+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/PropertyNames.java
@@ -65,10 +65,5 @@ public class PropertyNames {
   @Deprecated
   public static final String CONTAINER_USE_CURRENT_TIMESTAMP = "jib.container.useCurrentTimestamp";
 
-  @Deprecated public static final String EXTRA_DIRECTORY_PATH = "jib.extraDirectory.path";
-
-  @Deprecated
-  public static final String EXTRA_DIRECTORY_PERMISSIONS = "jib.extraDirectory.permissions";
-
   private PropertyNames() {}
 }